requirements.txt
sanic==20.9.0
python-binance==0.7.5
jinja2==2.11.2
